+++ changelog.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - data validation from the AllCohorts rule
 - move the schedule to the second row of the schedule page
 
-## v0.4.0 - Unreleased
+## v0.4.0 - 2025.03.08
 
 ### Added
 - OneClassAtATime rule to prevent cohorts from being scheduled for multiple classes at the same time